requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1

# Validation & settings
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Logging
structlog==24.1.0

# Dev / Testing
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
pytest-cov==5.0.0
